+++ Source/DafnyCore/AST/Grammar/ProgramParser.cs
@@ -12,17 +12,6 @@ using static Microsoft.Dafny.ParseErrors;
 
 namespace Microsoft.Dafny;
 
-public interface IFileSystem {
-  TextReader ReadFile(Uri uri);
-}
-
-public class OnDiskFileSystem : IFileSystem {
-  public static readonly IFileSystem Instance = new OnDiskFileSystem();
-  public TextReader ReadFile(Uri uri) {
-    return new StreamReader(File.Open(uri.LocalPath, FileMode.Open));
-  }
-}
-
 public record DfyParseResult(
   BatchErrorReporter ErrorReporter,
   FileModuleDefinition Module,
