+++ Source/DafnyCore/Generic/Node.cs
@@ -28,17 +28,7 @@ public interface IHasDocstring : INode {
   string GetDocstring(DafnyOptions options);
 }
 
-public class EmptyNode : Node {
-  public override RangeToken RangeToken { get; set; } = new(new Token(), new Token());
-  public override IToken Tok => new Token();
-  public override IEnumerable<Node> Children => Enumerable.Empty<Node>();
-  public override IEnumerable<Node> PreResolveChildren => Enumerable.Empty<Node>();
-}
-
 public abstract class Node : INode {
-
-
-
   public abstract IToken Tok { get; }
 
   /// <summary>
